+++ inferno-lsp/CHANGELOG.md
@@ -1,6 +1,15 @@
 # Revision History for inferno-lsp
 *Note*: we use https://pvp.haskell.org/ (MAJOR.MAJOR.MINOR.PATCH)
 
+## 0.1.7 -- 2023-07-11
+* Update inferno-types version
+
+## 0.1.6 -- 2023-07-11
+* Add a timeout to parseAndInfer
+
+## 0.1.5 -- 2023-06-26
+* Update inferno-vc version
+
 ## 0.1.4 -- 2023-06-19
 * Raise error if script evaluates to a function (and suggest adding input parameters instead)